Fixed makefile and comparison of signed number and unsigned in 2 functions

+++ ft_putull_base_fd.c
@@ -6,7 +6,7 @@
 /*   By: emajuri <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/28 09:58:30 by emajuri           #+#    #+#             */
-/*   Updated: 2022/11/22 13:53:34 by emajuri          ###   ########.fr       */
+/*   Updated: 2022/11/23 19:28:59 by emajuri          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,11 +17,11 @@ void	ft_putull_base_fd(unsigned long long n, int base, int fd)
 	static char	*str = "0123456789abcdefg";
 	if (base < 2 || base > 16)
 		return ;
-	if (n > (base - 1))
+	if (n > (unsigned int)(base - 1))
 	{
 		ft_putull_base_fd(n / base, base, fd);
 		ft_putull_base_fd(n % base, base,  fd);
 	}
-	if (n <= (base - 1) && n >= 0)
+	if (n <= (unsigned int)(base - 1) && n >= 0)
 		ft_putchar_fd(str[n], fd);
 }
